completing TODO 1

+++ src/DataTypes.java
@@ -4,7 +4,7 @@ public class DataTypes {
     // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(List<Integer> numbers) {
 
-        int s = 0;
+        long s = 0; // because the int datatype has a max value it can store
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
